+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - bass property added to Chord
 
+### Changed
+
+- Renamed Scale.degree to Scale.getPitchByDegree
+
 ### Removed
 
 - Scale.getIndex